working fine i think - 16 minutes everything. create cache(maybe need improve) can save cache and finalDic
need to work on parser and gui.
need to adjust to gui.

indexer has a function to save cacheWordsList in the project folder(for first run) after need to check that will always work.

postingIO and MERGER have a flag by indexer if stemmin is used and works on diff folder "stemPosting"